+++ invocations/analysis.py
@@ -7,8 +7,13 @@ from pathlib import Path
 
 from invoke import task, Collection
 
-from invocations import shell, config
-from invocations.util import REPO_ROOT, safely_load_config, remove_blank_lines
+from invocations import shell, config, docs
+from invocations.util import (
+    REPO_ROOT,
+    MAIN_PACKAGE,
+    safely_load_config,
+    remove_blank_lines,
+)
 
 try:
     from colorama import Fore, Style
@@ -25,6 +30,8 @@ ROBOT_TEST_SOURCE = Path("tests/robot")
 ROBOT_TEST_OUTPUT = Path("tests/results")
 ROBOT_TEST_RESOURCES = Path("tests/resources")
 
+MAIN_README = MAIN_PACKAGE / "README.rst"
+
 EXCLUDE_ROBOT_TESTS = ["skip"]
 if platform.system() == "Windows":
     EXCLUDE_ROBOT_TESTS.append("posix")
@@ -32,22 +39,58 @@ else:
     EXCLUDE_ROBOT_TESTS.append("windows")
 
 
-@task(config.install)
-def lint(ctx):
-    """Run format checks and static analysis"""
-    if getattr(ctx, "is_meta", False):
-        shell.invoke_each(ctx, "code.lint")
-    else:
-        flake8_config = Path(
-            safely_load_config(ctx, "ctx.linters.flake8", FLAKE8_CONFIG)
-        )
-        pylint_config = Path(
-            safely_load_config(ctx, "ctx.linters.pylint", PYLINT_CONFIG)
-        )
+@task(
+    config.install,
+    help={
+        "docstrings": "Also check docstring format.",
+        "all": "Run linting against all packages as well.",
+        "exit_on_failure": "Causes task to exit when checks fail. Used for publish and build.",
+    },
+)
+def lint(ctx, docstrings=False, all=False, exit_on_failure=False):
+    """Run format checks and static analysis. By default this task does
+    not exit when checks fail.
 
-        shell.poetry(ctx, "run black --diff --check src")
-        shell.poetry(ctx, f"run flake8 --config {flake8_config} src")
-        shell.poetry(ctx, f"run pylint --rcfile {pylint_config} src")
+    When ran at the meta package level, this task runs linters against
+    documentation and does a dummy build of the Sphinx docs. You can
+    also have it run the lint command for all packages by setting
+    ``--all``.
+    """
+    warn_setting = not exit_on_failure
+    flake8_config = Path(safely_load_config(ctx, "ctx.linters.flake8", FLAKE8_CONFIG))
+    pylint_config = Path(safely_load_config(ctx, "ctx.linters.pylint", PYLINT_CONFIG))
+    if docstrings:
+        ignore_codes_cmd = ""
+        all_docstrings_cmd = "--docstrings"
+    else:
+        ignore_codes_cmd = "--extend-ignore D,RST"
+        all_docstrings_cmd = ""
+    if getattr(ctx, "is_meta", False):
+        shell.poetry(
+            ctx,
+            f"run py -m sphinxlint {docs.DOCS_SOURCE_DIR} {MAIN_README}",
+            warn=warn_setting,
+        )
+        shell.sphinx(
+            ctx,
+            f"-b dummy -a -n --keep-going {docs.DOCS_SOURCE_DIR} {docs.DOCS_BUILD_DIR}",
+            warn=warn_setting,
+        )
+        shell.poetry(
+            ctx,
+            f"run flake8 --config {flake8_config} {docs.DOCS_SOURCE_DIR}",
+            warn=warn_setting,
+        )
+        if all:
+            shell.invoke_each(ctx, f"code.lint {all_docstrings_cmd}")
+    else:
+        shell.poetry(ctx, "run black --diff --check src", warn=warn_setting)
+        shell.poetry(
+            ctx,
+            f"run flake8 --config {flake8_config} {ignore_codes_cmd} src",
+            warn=warn_setting,
+        )
+        shell.poetry(ctx, f"run pylint --rcfile {pylint_config} src", warn=warn_setting)
 
 
 @task(config.install, aliases=["pretty"])