+++ packages/main/src/RPA/MSGraph.py
@@ -190,14 +190,28 @@ class MSGraph:
         else:
             return storage.get_default_drive()
 
-    def _parse_drive_param(self, drive: DriveType) -> drive_module.Drive:
+    def _parse_drive_param(
+        self,
+        drive: Optional[DriveType] = None,
+        site: Optional[sharepoint.Site] = None,
+        resource: Optional[str] = None,
+    ) -> drive_module.Drive:
         """Helper function which parses the drive parameter given
-        from a robot.
+        from a robot. If you pass a sharepoint site, it will use
+        that to retrieve the default drive rather than OneDrive.
         """
         if isinstance(drive, drive_module.Drive):
             return drive
+        elif isinstance(drive, str):
+            return (
+                site.get_document_library(drive)
+                if site is not None
+                else self.get_drive_instance(resource, drive)
+            )
+        elif site is not None:
+            return site.get_default_document_library()
         else:
-            return self.get_drive_instance(drive)
+            return self.get_drive_instance(resource)
 
     @keyword
     def get_folder_instance(
@@ -312,16 +326,6 @@ class MSGraph:
         # Method download_contents has no return value.
         folder.download_contents(to_folder=downloaded_folder)
         return downloaded_folder
-
-    def _get_sharepoint_drive(
-        self, site: sharepoint.Site, drive: Optional[DriveType] = None
-    ) -> drive_module.Drive:
-        """Returns the specified SharePoint drive if any or the default one if none."""
-        drive = self._parse_drive_param(drive)
-        if drive:
-            return site.get_document_library(drive.object_id)
-        else:
-            return site.get_default_document_library()
 
     def _sharepoint_items_into_dict_list(
         self, items_instance: List[sharepoint.SharepointListItem]
@@ -579,17 +583,19 @@ class MSGraph:
         target_folder: Union[drive_module.Folder, str, None] = None,
         include_folders: Optional[bool] = False,
         resource: Optional[str] = None,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
     ) -> List[drive_module.DriveItem]:
         """Returns a list of files from the specified OneDrive folder.
 
         The files returned are ``DriveItem`` objects and they have additional
         properties that can be accessed with dot-notation.
 
-        :param target_folder: Path of the folder in Onedrive_module.
+        :param target_folder: Path of the folder in Onedrive. If none is
+         provided, will return files from the root of Onedrive.
         :param include_folders: Boolean indicating if should return folders as well.
         :param resource: Name of the resource if not using default.
-        :param drive_id: Drive ID if not using default.
+        :param drive: A ``Drive`` object or Drive ID as a string. If not
+         provided, will use the default drive.
         :return: List of DriveItems in the folder.
 
         .. code-block: robotframework
@@ -612,7 +618,7 @@ class MSGraph:
                 END
         """
         self._require_authentication()
-        drive_instance = self.get_drive_instance(resource, drive_id)
+        drive_instance = self._parse_drive_param(drive, resource=resource)
         folder = self.get_folder_instance(drive_instance, target_folder)
         items = folder.get_items()
         if include_folders:
@@ -626,9 +632,9 @@ class MSGraph:
         to_path: Optional[PathType] = None,
         name: Optional[str] = None,
         resource: Optional[str] = None,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
     ) -> Path:
-        """Downloads a file from Onedrive_module.
+        """Downloads a file from Onedrive.
 
         The downloaded file will be saved to a local path.
 
@@ -637,7 +643,8 @@ class MSGraph:
          defaults to the current directory.
         :param name: New name for the downloaded file, with or without extension.
         :param resource: Name of the resource if not using default.
-        :param drive_id: Drive ID if not using default.
+        :param drive: A ``Drive`` object or Drive ID as a string. If not
+         provided, will use the default drive.
         :return: Path to the downloaded file.
 
         .. code-block: robotframework
@@ -656,7 +663,7 @@ class MSGraph:
                 ...    Report.pdf
         """
         self._require_authentication()
-        drive_instance = self.get_drive_instance(resource, drive_id)
+        drive_instance = self._parse_drive_param(drive, resource=resource)
         file_instance = self.get_file_instance(drive_instance, target_file)
         return self._download_file(file_instance, to_path, name)
 
@@ -666,7 +673,7 @@ class MSGraph:
         target_folder: Union[drive_module.Folder, str],
         to_path: Optional[PathType] = None,
         resource: Optional[str] = None,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
     ) -> Path:
         """Downloads a folder from OneDrive with all of its contents,
         including subfolders.
@@ -678,7 +685,8 @@ class MSGraph:
         :param to_path: Destination folder where the download will be saved to,
          defaults to the current directory.
         :param resource: Name of the resource if not using default.
-        :param drive_id: Drive ID if not using default.
+        :param drive: A ``Drive`` object or Drive ID as a string. If not
+         provided, will use the default drive.
         :return: Path to the downloaded folder.
 
         .. code-block: robotframework
@@ -695,7 +703,7 @@ class MSGraph:
                 ...    /path/to/local/folder
         """
         self._require_authentication()
-        drive_instance = self.get_drive_instance(resource, drive_id)
+        drive_instance = self._parse_drive_param(drive, resource=resource)
         folder_instance = self.get_folder_instance(drive_instance, target_folder)
         return self._download_folder(folder_instance, to_path)
 
@@ -706,7 +714,7 @@ class MSGraph:
         target_folder: Optional[Union[drive_module.Folder, str]] = None,
         include_folders: Optional[bool] = False,
         resource: Optional[str] = None,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
     ) -> List[drive_module.DriveItem]:
         # pylint: disable=anomalous-backslash-in-string
         """Returns a list of files found in OneDrive based on the search string.
@@ -718,13 +726,14 @@ class MSGraph:
         properties that can be accessed with dot-notation, see
         \`List Files In Onedrive Folder\` for details.
 
-        :param search_string: String used to search for file in Onedrive_module.
+        :param search_string: String used to search for file in Onedrive.
          Values may be matched across several fields including filename,
          metadata, and file content.
         :param target_folder: Folder where to search for files.
         :param include_folders: Boolean indicating if should return folders as well.
         :param resource: Name of the resource if not using default.
-        :param drive_id: Drive ID if not using default.
+        :param drive: A ``Drive`` object or Drive ID as a string. If not
+         provided, will use the default drive.
         :return: List of DriveItems found based on the search string.
 
         .. code-block: robotframework
@@ -734,7 +743,7 @@ class MSGraph:
                 ${files}=    Find Onedrive File    Report.xlsx
         """  # noqa: W605
         self._require_authentication()
-        drive_instance = self.get_drive_instance(resource, drive_id)
+        drive_instance = self._parse_drive_param(drive, resource=resource)
         if target_folder:
             folder = self.get_folder_instance(drive_instance, target_folder)
             items = folder.search(search_string)
@@ -759,6 +768,7 @@ class MSGraph:
         :param to_path: Destination folder of the downloaded file,
          defaults to the current directory.
         :param name: New name for the downloaded file, with or without extension.
+        :return: The Path to the downloaded file.
 
         .. code-block: robotframework
 
@@ -792,7 +802,7 @@ class MSGraph:
         file_path: str,
         target_folder: Optional[Union[drive_module.Folder, str]] = None,
         resource: Optional[str] = None,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
     ) -> drive_module.DriveItem:
         # pylint: disable=anomalous-backslash-in-string
         """Uploads a file to the specified OneDrive folder.
@@ -802,9 +812,11 @@ class MSGraph:
         \`List Files In Onedrive Folder\` for details.
 
         :param file_path: Path of the local file being uploaded.
-        :param target_folder: Path of the folder in Onedrive_module.
+        :param target_folder: Path of the folder in Onedrive.
         :param resource: Name of the resource if not using default.
-        :param drive_id: Drive ID if not using default.
+        :param drive: A ``Drive`` object or Drive ID as a string. If not
+         provided, will use the default drive.
+        :return: The newly created DriveItem.
 
         .. code-block: robotframework
 
@@ -815,7 +827,7 @@ class MSGraph:
                 ...    /path/to/folder
         """  # noqa: W605
         self._require_authentication()
-        drive_instance = self.get_drive_instance(resource, drive_id)
+        drive_instance = self._parse_drive_param(drive, resource=resource)
         folder = self.get_folder_instance(drive_instance, target_folder)
         return folder.upload_file(item=file_path)
 
@@ -971,13 +983,14 @@ class MSGraph:
         self,
         site: sharepoint.Site,
         include_folders: Optional[bool] = False,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
+        target_folder: Union[drive_module.Folder, str, None] = None,
     ) -> List[drive_module.DriveItem]:
         # pylint: disable=anomalous-backslash-in-string
-        """List files in the SharePoint Site drive_module.
+        """List files in the SharePoint Site drive.
 
-        If the ``drive_id`` is not set, the default Document Library
-        will be used. The ``drive_id`` can be obtained from the keyword
+        If the ``drive`` is not set, the default Document Library
+        will be used. The ``drive`` can be obtained from the keyword
         \`List Sharepoint Site Drives\`.
 
         The files returned are ``DriveItem`` objects and they have additional
@@ -987,8 +1000,11 @@ class MSGraph:
         :param site: Site instance obtained from \`Get Sharepoint Site\`.
         :param include_folders: Boolean indicating if should return folders
          as well.
-        :param drive_id: ID of the desired drive_module.
-        :return: List of DriveItems present in the Site drive_module.
+        :param drive: A ``Drive`` object or Drive ID as a string. If not
+         provided, will use the default drive.
+        :param target_folder: Path of the folder in the Sharepoint drive.
+         If none is provided, will return files from the root of the drive.
+        :return: List of DriveItems present in the Site drive.
 
         .. code-block: robotframework
 
@@ -998,8 +1014,8 @@ class MSGraph:
                 ${files}    List Files In Sharepoint Site Drive    ${site}
         """  # noqa: W605
         self._require_authentication()
-        sp_drive = self._get_sharepoint_drive(site, drive_id)
-        folder = self.get_folder_instance(sp_drive)
+        drive = self._parse_drive_param(drive, site=site)
+        folder = self.get_folder_instance(drive, target_folder)
         items = folder.get_items()
         if include_folders:
             return items
@@ -1012,7 +1028,7 @@ class MSGraph:
         site: sharepoint.Site,
         to_path: Optional[PathType] = None,
         name: Optional[str] = None,
-        drive_id: Optional[str] = None,
+        drive: Optional[DriveType] = None,
     ) -> Path:
         # pylint: disable=anomalous-backslash-in-string
         """Downloads file from SharePoint.
@@ -1024,7 +1040,7 @@ class MSGraph:
         :param to_path: Destination folder of the downloaded file,
          defaults to the current directory.
         :param name: New name for the downloaded file, with or without extension.
-        :param drive_id: Drive ID if not using default.
+        :param drive: A ``Drive`` object or a Drive ID as a string.
         :return: Path to the downloaded file.
 
         .. code-block: robotframework
@@ -1045,6 +1061,6 @@ class MSGraph:
                 ...    Report.pdf
         """  # noqa: W605
         self._require_authentication()
-        sp_drive = self._get_sharepoint_drive(site, drive_id)
-        file_instance = self.get_file_instance(sp_drive, target_file)
+        drive = self._parse_drive_param(drive, site=site)
+        file_instance = self.get_file_instance(drive, target_file)
         return self._download_file(file_instance, to_path, name)